Remove -Wno-sign-compare and fix all resulting warnings
Fix 13 sign-compare warnings in hand-written code by correcting
variable types to match their unsigned comparands (sizeof, strlen,
LENGTH_OF). For two functions where the loop counter was shared across
int-bounded and size_t-bounded loops, add a separate size_t variable.
Suppress the 9 unfixable warnings in flex-generated runtime code via a
targeted FLEX_PROGS rule in gkdict/makefile.

+++ src/gener/conjsys.c
@@ -273,8 +273,8 @@ int noppart(char *s)
 {
 	char buf1[256];
 	char buf2[256];
-	int i;
-	
+	size_t i;
+
 	if(is_empty(s) ) return(1);
 	
 	Xstrncpy(buf1,s,sizeof buf1);
